requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
beautifulsoup4
pytest
Pillow

